+++ EssentialFeediOS/Feed Image Loader/FeedImageDataLoader.swift
@@ -1,0 +1,19 @@
+//
+//  FeedImageDataLoader.swift
+//  EssentialFeediOS
+//
+//  Created by Christos Tzimas on 28/5/25.
+//
+
+import Foundation
+
+public protocol FeedImageDataLoaderTask {
+    func cancel()
+}
+
+public protocol FeedImageDataLoader {
+    
+    typealias Result = Swift.Result<Data, Error>
+
+    func loadImageData(from url: URL, completion: @escaping (Result) -> Void) -> FeedImageDataLoaderTask
+}
